+++ noodles-cram/src/async/reader/data_container/compression_header/data_series_encoding_map.rs
@@ -87,6 +87,7 @@ where
             DataSeries::MappingQualities => builder.set_mapping_qualities_encoding(encoding),
             DataSeries::Bases => builder.set_bases_encoding(encoding),
             DataSeries::QualityScores => builder.set_quality_scores_encoding(encoding),
+            DataSeries::ReservedTc | DataSeries::ReservedTn => builder,
         }
     }